Create pinpoint.py
Source Code Description: 
pinpoint.py is the latest revision of the pinpoint source code. It performs a series of spiral samples starting at 0.3m up to 0.7m with 0.2 increments. The sampling space if reduced using the maximum concentration location as the epicenter to define a new smaller spiral sampling area. 
Known Issues: 
The drone seems to experience difficulty in maintaining its altitude when it is flying above 0.7m. It seems a firmware issue. 
When drone goes to areas that have not been visited, there is no mechanism to resample to make certain the reduced sampling spaces are evaluated from visited areas history.